+++ src/ds5d.cpp
@@ -140,7 +140,7 @@ namespace rsimpl
             switch (options[i])
             {
             case RS_OPTION_DS5_LASER_POWER:     ds5::set_laser_power(get_device(), static_cast<uint8_t>(values[i])); break;
-            case RS_OPTION_R200_LR_EXPOSURE:    ds5::set_lr_exposure(get_device(), static_cast<uint32_t>(values[i])); break;
+            case RS_OPTION_R200_LR_EXPOSURE:    ds5::set_lr_exposure(get_device(), static_cast<uint16_t>(values[i])); break;
 
             default: base_opt.push_back(options[i]); base_opt_val.push_back(values[i]); break;
             }
